Move DocumentMappingService down to Workspaces
