+++ src/Razor/src/Microsoft.AspNetCore.Razor.LanguageServer/Resources/xlf/SR.zh-Hans.xlf
@@ -109,12 +109,7 @@
       </trans-unit>
       <trans-unit id="Unknown_RazorMetaCode">
         <source>Attempted to visit a RazorMetaCode other than '{' or '}'.</source>
-        <target state="translated">尝试访问了除 "{" 或 "}" 之外的 RazorMetaCode。</target>
-        <note />
-      </trans-unit>
-      <trans-unit id="Unknown_mapping_behavior">
-        <source>Unknown mapping behavior</source>
-        <target state="translated">未知映射行为</target>
+        <target state="new">Attempted to visit a RazorMetaCode other than '{' or '}'.</target>
         <note />
       </trans-unit>
       <trans-unit id="Version_Should_Not_Be_Null">
