+++ src/Razor/src/Microsoft.CodeAnalysis.Razor.Workspaces/Resources/SR.resx
@@ -136,4 +136,7 @@
   <data name="PositionLine_Outside_Range" xml:space="preserve">
     <value>Line '{0}' outside of the {1} range of '{2}' was queried. The document may not be up to date.</value>
   </data>
+  <data name="Unknown_mapping_behavior" xml:space="preserve">
+    <value>Unknown mapping behavior</value>
+  </data>
 </root>
